Tabbed Windows
This gives the option to open all of the window dialogs as tabbed resources. The plugin writer and plugin console tab could be merged into a single pane at a later date. The error dialog might be over-kill and should remain an external window.

This is mostly an experiment for better UIs, because of this everything is optional and can be turned off within the configuration class