Continue Decompiling On Any Parsing Errors (Except IO)
More of a pre-caution - if parsing fails we should still attempt to display the decompiled code. I've encountered a few errors and manually added them in, but I'm worried I missed one. To fix that I figure we just catch everything that isn't an IO Error.

IO-Errors are probably user related (lack of space, insufficient permissions, etc) - So these errors we should forward to the user, the rest we can silence for developer eyes